SPQR: to v4.0, int32 demos, flop counts revised, docs

+++ GPUQREngine/Include/GPUQREngine_Scheduler.hpp
@@ -157,7 +157,7 @@ public:
     /* Stats */
     float kernelTime;
     Int numKernelLaunches;
-    Int gpuFlops;
+    int64_t gpuFlops;
 
 #ifdef GPUQRENGINE_RENDER
     /* Debug stuff */
